Add function humanMove&computerMove
humanMove - Reads the movement specified by the user. Returns the user's move. computerMove - Artificial intelligence tries to find the best move. First - tries to find a move that will win the game. Then try to block the move that will give victory to the human. Then choose the best possible move (Centre of the board > Corners of the board > Rest).